refactor to verify email hook
